requirements.txt
async-timeout==4.0.2
bcrypt==4.0.1
cffi==1.15.1
click==8.0.4
cryptography==40.0.2
dataclasses==0.8
Flask==2.0.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
paramiko==3.2.0
pycparser==2.21
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.7
PyYAML==6.0
redis==4.3.6
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
pycrypto==2.6.1
